+++ tests/project_items/combiner/test_Combiner.py
@@ -60,7 +60,7 @@ class TestCombiner(unittest.TestCase):
     def test_item_dict(self):
         """Tests Item dictionary creation."""
         d = self.combiner.item_dict()
-        a = ["type", "short name", "description", "x", "y", "cancel_on_error"]
+        a = ["type", "description", "x", "y", "cancel_on_error"]
         for k in a:
             self.assertTrue(k in d, f"Key '{k}' not in dict {d}")