+++ CHANGELOG.md
@@ -39,6 +39,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
   each individual item because the exact order is not know before the Execute button is actually clicked.
   The number still indicates the execution order but may show the same numbers for items in different parallel
   branches.
+- Project.json file format has been upgraded to version 2. Version 1 project.json files are upgraded to version 2
+  automatically when a project is opened. 
 
 
 ### Deprecated